Added part of nav. ( clic arrow, change article p)
